Add load-from-loader-memory code path for rda memory loads
